@@ -4,4 +4,4 @@
 # If you change this key, all old signed cookies will become invalid!
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-Depot::Application.config.secret_token = '8e9c938524f6d81ff4ba5be6fcbdac76b758e0a25933cfb956353554e918dcbae7dfe5424b0e09b2d220e5868d0a08df88c4174d0301ce3f14e2d25c5a99895a'
+Depot::Application.config.secret_token = '50f8522a48d6110f939c0a98530e1d511901ac769a65eec352046a9afa0df90e5c745a461b12d603773e008baf9bece7f5c17837aa4ed8f22e8b3219453aab06'
